+++ pykafka/exceptions.py
@@ -40,11 +40,6 @@ class ConsumerStoppedException(KafkaException):
     pass
 
 
-class NoPartitionsForConsumerException(ConsumerStoppedException):
-    """Indicates that this consumer is stopped because it assigned itself zero partitions"""
-    pass
-
-
 class NoMessagesConsumedError(KafkaException):
     """Indicates that no messages were returned from a MessageSet"""
     pass
@@ -149,10 +144,11 @@ class OffsetMetadataTooLarge(ProtocolClientError):
     ERROR_CODE = 12
 
 
-class OffsetsLoadInProgress(ProtocolClientError):
+class GroupLoadInProgress(ProtocolClientError):
     """The broker returns this error code for an offset fetch request if it is
         still loading offsets (after a leader change for that offsets topic
-        partition).
+        partition), or in response to group membership requests (such as
+        heartbeats) when group metadata is being loaded by the coordinator.
     """
     ERROR_CODE = 14
 
@@ -183,7 +179,7 @@ ERROR_CODES = dict(
                 RequestTimedOut,
                 MessageSizeTooLarge,
                 OffsetMetadataTooLarge,
-                OffsetsLoadInProgress,
+                GroupLoadInProgress,
                 ConsumerCoordinatorNotAvailable,
                 NotCoordinatorForConsumer)
 )